+++ back/.gitignore
@@ -1,3 +1,4 @@
+/dist/
 /node_modules/
 /dist/bundle.js
 /yarn-error.log
