+++ back/src/Controller/IoSocketController.ts
@@ -1,7 +1,7 @@
 import socketIO = require('socket.io');
 import {Socket} from "socket.io";
 import * as http from "http";
-import {JoinRoomMessage} from "@Model/Websocket/JoinRoomMessage";
+import {MessageUserPosition} from "@Model/Websocket/MessageUserPosition";
 
 export class IoSocketController{
     Io: socketIO.Server;
@@ -21,8 +21,14 @@ export class IoSocketController{
                     positionXUser: user x position map
                     positionYUser: user y position on map
             */
-            socket.on('join-room', (message : JoinRoomMessage) => {
+            socket.on('join-room', (message : MessageUserPosition) => {
                 socket.join(message.roomId);
+                // sending to all clients in room except sender
+                socket.to(message.roomId).emit('join-room', message.toString());
+            });
+
+            socket.on('user-position', (message : MessageUserPosition) => {
+                // sending to all clients in room except sender
                 socket.to(message.roomId).emit('join-room', message.toString());
             });
         });
